list: add slist.h for singly-linked list (WIP)
AFAIK there's no "proper" (as in "same API as ccan/list") singly-linked
list implementation in CCAN.

The appended is obviously incomplete (e.g. no debug, no deletions) but as is
it gives me a ~15% speedup when doing a BFS traversal of a large graph, whose
edges are represented as an adjacency list.

- Should we have something like this? I tried adding a for_each macro to
  lqueue but that was just too ugly to live. Yes, one could just use *next
  pointers in the original code, but then there's no list_for_each, and
  no trivial update path to a doubly-linked list should the need arise.

- If so, should it be a separate module, a submodule under list, or just
  a header next to list.h?
  I'm just dropping slist.h here due to laziness; I like the idea of having
  both singly and doubly-linked list implementations under list/ though,
  so that they can share code and are less likely be overlooked by users.

Signed-off-by: Emilio G. Cota <cota@braap.org>